Create 3DParticleSimulator
This is a 3D particle simulator that portrays the position of a particle in x,y,z cartesian coordinates. The particle's movement is influenced by an electric field, magnetic field, and initial velocity specified by user input. Further, the position is dependent upon the particles charge and mass. The sliders control the magnitudes of each value in the different directions.